Fix bug - wrong type used in memcpy()

+++ Source/Utils/TyphoonIntrospectionUtils.m
@@ -39,7 +39,7 @@ BOOL TyphoonIsInvalidClassName(NSString *className);
             return (NULL);
         }
 
-        int len = (int) (e - attributes);
+        NSUInteger len = (NSUInteger) (e - attributes);
         memcpy( buffer, attributes, len );
         buffer[len] = '\0';
